+++ src/main/java/com/b1n_ry/yigd/Yigd.java
@@ -222,6 +222,9 @@ public class Yigd implements ModInitializer, DedicatedServerModInitializer {
             apiMods.add(new OriginsCompat());
             miscCompatMods.add("apoli");
         }
+        if (FabricLoader.getInstance().isModLoaded("numismatic-overhaul")) {
+            apiMods.add(new NumismaticOverhaulCompat());
+        }
         apiMods.addAll(FabricLoader.getInstance().getEntrypoints("yigd", YigdApi.class));
 
         if (FabricLoader.getInstance().isModLoaded("flan")) {